requirements.txt
pandas>=2.0
requests>=2.31
pydantic>=2.6
PyYAML>=6.0
rdflib>=7.0
pyshacl>=0.24
typer>=0.12
streamlit>=1.32
-e .
plotly>=5.20
duckduckgo-search>=6.2.6
readability-lxml>=0.8.1
beautifulsoup4>=4.12
openpyxl>=3.1
openai>=1.0
anthropic>=0.8
-e .
